Updated Bottom Navigation Menu
