+++ lib/features/authentication/screens/login/login.dart
@@ -10,7 +10,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:iconsax/iconsax.dart';
 
-import '../../../../navigation_mnu.dart';
+import '../../../../navigation_menu.dart';
 import '../password_configuration/forgot_password.dart';
 import '../signup/signup.dart';
